test: verify no-mistakes QA pipeline (#1)
* test(pipeline): trigger no-mistakes QA pipeline

- Add javadoc to ApiApplication to verify pipeline triggers
- Testing: rebase check, backend compile, frontend build, docs validation

Refs: test/no-mistakes-pipeline

* fix(ci): use mvn instead of mvnw, add npm fallback

- Replace ./mvnw with mvn (GitHub Actions runners have Maven pre-installed)
- Add Maven cache to setup-java
- Add npm cache to setup-node
- Fallback to npm install when package-lock.json is missing
- Remove spotless check since it's not configured in pom.xml

Refs: test/no-mistakes-pipeline

* fix(ci): fix Java compilation error and frontend caching

- MediaService.java: make loop index effectively final for lambda (line 112)
- Remove npm cache config since package-lock.json doesn't exist yet

Refs: test/no-mistakes-pipeline

* fix(ci): resolve frontend and backend compilation errors

Backend:
- Remove precision/scale from Double columns in ListingLocation
  (only valid for BigDecimal, causes Hibernate 'scale has no meaning' error)

Frontend:
- Rename 'quote' property to 'priceQuote' in ListingDetailComponent
  to avoid Angular strict template compilation issue with shorthand property
- Fix maxNights type in ListingEditComponent (use ?? undefined instead of || null)
- Fix import paths in SearchBarComponent (../listings instead of ../../listings)

Refs: test/no-mistakes-pipeline

* fix(ci): resolve TypeScript and database schema issues

Backend:
- Change latitude/longitude from NUMERIC to DOUBLE PRECISION in migration V2
- Add DockerImageName.asCompatibleSubstituteFor to Testcontainers config
  (fixes postgis image compatibility check)

Frontend:
- Replace Quote interface with individual primitive properties in ListingDetail
  (avoids Angular strict template type narrowing issues)
- Fix maxNights form control type in ListingEdit
- Remove unused imports from SearchBar

Refs: test/no-mistakes-pipeline

* fix(ci): extract Angular template to HTML file and fix MockMvc paths

Frontend:
- Extract listing-detail template to external HTML file
  (fixes TypeScript template literal interpolation conflict with dollar-curlybrace syntax)

Backend:
- Remove /api prefix from MockMvc test requests
  (context-path is handled automatically by MockMvc)

Refs: test/no-mistakes-pipeline

* fix(ci): resolve backend test failures and Angular template interpolation errors

Backend:
- Use valid UUID for local/ci JWT sub claim to fix UUID parsing in controllers
- Fix PostgreSQL lower(bytea) type error by pre-building location pattern in Java
  instead of using CONCAT with nullable parameter in JPQL

Frontend:
- Extract inline templates to external .html files to avoid TypeScript template
  literal interpolation errors with dollar-curly-brace syntax
- Fix relative import paths in host and listings components
- Add missing ListingCardDto export to listing.model
- Fix implicit any types and form.value unknown types in listing create/edit

* fix(ci): seed local test user for integration tests

* fix(ci): use Hibernate JSON type mapping for PostgreSQL jsonb columns

* fix(ci): add PR write permissions for qa-gate comment step

---------

+++ apps/api-spring/src/main/resources/db/migration/V2__listings_domain.sql
@@ -41,8 +41,8 @@ CREATE TABLE IF NOT EXISTS listing_locations (
     state VARCHAR(128),
     postal_code VARCHAR(32),
     country VARCHAR(128) NOT NULL,
-    latitude NUMERIC(10, 8),
-    longitude NUMERIC(11, 8)
+    latitude DOUBLE PRECISION,
+    longitude DOUBLE PRECISION
 );
 
 CREATE TABLE IF NOT EXISTS listing_policies (
